+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-03T05:16:03.721Z",
+    "2026-02-03T08:30:30.445Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {